src/nightly_build/requirements.txt
git-remote-codecommit==1.15.1
gitdb==4.0.7
GitPython==3.1.18
